refactor: move test suite actions to dedicated route files
