+++ packages/insomnia/src/ui/routes/$organizationId.project.$projectId.workspace.$workspaceId.test-suite.$testSuiteId.delete.tsx
@@ -1,0 +1,22 @@
+import { type ActionFunctionArgs, redirect } from 'react-router';
+
+import * as models from '../../models';
+import { invariant } from '../../utils/invariant';
+import { SegmentEvent } from '../analytics';
+
+export async function action({ params }: ActionFunctionArgs) {
+  const { organizationId, workspaceId, projectId, testSuiteId } = params;
+  invariant(typeof testSuiteId === 'string', 'Test Suite ID is required');
+  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
+  invariant(typeof projectId === 'string', 'Project ID is required');
+
+  const unitTestSuite = await models.unitTestSuite.getById(testSuiteId);
+
+  invariant(unitTestSuite, 'Test Suite not found');
+
+  await models.unitTestSuite.remove(unitTestSuite);
+
+  window.main.trackSegmentEvent({ event: SegmentEvent.testSuiteDelete });
+
+  return redirect(`/organization/${organizationId}/project/${projectId}/workspace/${workspaceId}/test`);
+}
