+++ packages/insomnia/src/ui/routes/$organizationId.project.$projectId.workspace.$workspaceId.test-suite.$testSuiteId.update.tsx
@@ -1,0 +1,25 @@
+import { type ActionFunctionArgs } from 'react-router';
+
+import { database } from '../../common/database';
+import * as models from '../../models';
+import type { UnitTestSuite } from '../../models/unit-test-suite';
+import { invariant } from '../../utils/invariant';
+
+export async function action({ request, params }: ActionFunctionArgs) {
+  const { workspaceId, projectId, testSuiteId } = params;
+  invariant(typeof testSuiteId === 'string', 'Test Suite ID is required');
+  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
+  invariant(typeof projectId === 'string', 'Project ID is required');
+
+  const data = (await request.json()) as Partial<UnitTestSuite>;
+
+  const unitTestSuite = await database.getWhere<UnitTestSuite>(models.unitTestSuite.type, {
+    _id: testSuiteId,
+  });
+
+  invariant(unitTestSuite, 'Test Suite not found');
+
+  await models.unitTestSuite.update(unitTestSuite, data);
+
+  return null;
+}
