+++ packages/insomnia/src/ui/routes/actions.tsx
@@ -18,7 +18,6 @@ import { isRequest, type Request } from '../../models/request';
 import { isRequestGroup, isRequestGroupId } from '../../models/request-group';
 import { isRequestGroupMeta } from '../../models/request-group-meta';
 import type { UnitTest } from '../../models/unit-test';
-import type { UnitTestSuite } from '../../models/unit-test-suite';
 import { getSendRequestCallback } from '../../network/unit-test-feature';
 import { insomniaFetch } from '../../ui/insomniaFetch';
 import { invariant } from '../../utils/invariant';
@@ -39,146 +38,7 @@ export function safeToUseInsomniaFileNameWithExt(fileName: string) {
   return `${safeToUseInsomniaFileName(fileName)}.yaml`;
 }
 
-// Test Suite
-
-export const deleteTestSuiteAction: ActionFunction = async ({ params }) => {
-  const { organizationId, workspaceId, projectId, testSuiteId } = params;
-  invariant(typeof testSuiteId === 'string', 'Test Suite ID is required');
-  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
-  invariant(typeof projectId === 'string', 'Project ID is required');
-
-  const unitTestSuite = await models.unitTestSuite.getById(testSuiteId);
-
-  invariant(unitTestSuite, 'Test Suite not found');
-
-  await models.unitTestSuite.remove(unitTestSuite);
-
-  window.main.trackSegmentEvent({ event: SegmentEvent.testSuiteDelete });
-
-  return redirect(`/organization/${organizationId}/project/${projectId}/workspace/${workspaceId}/test`);
-};
-
-export const runAllTestsAction: ActionFunction = async ({ params }) => {
-  const { organizationId, projectId, workspaceId, testSuiteId } = params;
-  invariant(typeof projectId === 'string', 'Project ID is required');
-  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
-  invariant(typeof testSuiteId === 'string', 'Test Suite ID is required');
-
-  const unitTests = await database.find<UnitTest>(models.unitTest.type, { parentId: testSuiteId }, { metaSortKey: 1 });
-  invariant(unitTests, 'No unit tests found');
-
-  const tests: Test[] = unitTests
-    .filter(t => t !== null)
-    .map(t => ({
-      name: t.name,
-      code: t.code,
-      defaultRequestId: t.requestId,
-    }));
-
-  const src = generate([{ name: 'My Suite', suites: [], tests }]);
-
-  const sendRequest = getSendRequestCallback();
-
-  let results: TestResults = {
-    failures: [],
-    passes: [],
-    pending: [],
-    stats: {
-      suites: 0,
-      tests: 0,
-      passes: 0,
-      pending: 0,
-      failures: 0,
-      start: undefined,
-      end: undefined,
-      duration: undefined,
-    },
-    tests: [],
-  };
-
-  try {
-    results = await runTests(src, { sendRequest });
-    const testResult = await models.unitTestResult.create({
-      results,
-      parentId: workspaceId,
-    });
-    window.main.trackSegmentEvent({ event: SegmentEvent.unitTestRunAll, properties: { organizationId, projectId } });
-
-    return redirect(
-      `/organization/${organizationId}/project/${projectId}/workspace/${workspaceId}/test/test-suite/${testSuiteId}/test-result/${testResult._id}`,
-    );
-  } catch (err) {
-    // create a result manually so that it can be displayed in the UI
-    results.stats.failures = 1;
-    results.stats.tests = 1;
-    results.tests.push({
-      currentRetry: 0,
-      duration: 0,
-      err: {
-        actual: undefined,
-        expected: undefined,
-        message: err.toString(),
-        multiple: [],
-        operator: undefined,
-        showDiff: false,
-        stack: '',
-      },
-      file: '',
-      fullTitle: 'Test Error',
-      id: '',
-      title: 'Test Error',
-    });
-    const testResult = await models.unitTestResult.create({
-      results,
-      parentId: workspaceId,
-    });
-    window.main.trackSegmentEvent({ event: SegmentEvent.unitTestRunAll, properties: { organizationId, projectId } });
-
-    return redirect(
-      `/organization/${organizationId}/project/${projectId}/workspace/${workspaceId}/test/test-suite/${testSuiteId}/test-result/${testResult._id}`,
-    );
-  }
-};
-
-export const updateTestSuiteAction: ActionFunction = async ({ request, params }) => {
-  const { workspaceId, projectId, testSuiteId } = params;
-  invariant(typeof testSuiteId === 'string', 'Test Suite ID is required');
-  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
-  invariant(typeof projectId === 'string', 'Project ID is required');
-
-  const data = (await request.json()) as Partial<UnitTestSuite>;
-
-  const unitTestSuite = await database.getWhere<UnitTestSuite>(models.unitTestSuite.type, {
-    _id: testSuiteId,
-  });
-
-  invariant(unitTestSuite, 'Test Suite not found');
-
-  await models.unitTestSuite.update(unitTestSuite, data);
-
-  return null;
-};
-
 // Unit Test
-export const createNewTestAction: ActionFunction = async ({ request, params }) => {
-  const { testSuiteId } = params;
-  invariant(typeof testSuiteId === 'string', 'Test Suite ID is required');
-  const formData = await request.formData();
-
-  const name = formData.get('name');
-  invariant(typeof name === 'string', 'Name is required');
-
-  await models.unitTest.create({
-    parentId: testSuiteId,
-    code: `const response1 = await insomnia.send();
-expect(response1.status).to.equal(200);`,
-    name,
-  });
-
-  window.main.trackSegmentEvent({ event: SegmentEvent.unitTestCreate });
-
-  return null;
-};
 
 export const deleteTestAction: ActionFunction = async ({ params }) => {
   const { testId } = params;
